Revert "refactor(suite-sync): rename reducer file to slice"
This reverts commit e3d041be61b375cc82e7916b1a0ae362d9e7df94.
