+++ suite-common/suite-sync/src/createTurnOffSuiteSync.ts
@@ -8,7 +8,7 @@ import {
 import { StaticSessionId } from '@trezor/connect';
 
 import { clearAll } from './data/suiteSyncDataReducer';
-import { updateSuiteSyncEnabled } from './suiteSyncSlice';
+import { updateSuiteSyncEnabled } from './suiteSyncReducer';
 
 export type CreateTurnOffSuiteSyncDeps = {
     getIsSuiteSyncEnabled: () => boolean;
